inductor_config_logging: Don't drop keys
This bit me while I was trying to debug some trace issues.
In general this config is already quite large when dumping, so adding
more fields doesn't make it significantly worse.

Also a number of the items we are type checking for (except the test
configs), don't even show up. Primarily this will help us when debugging
rocm, halide, and trace configs.

ghstack-source-id: bf25ac5d83310e9f9a2da43f7ab42b142920b540
Pull Request resolved: https://github.com/pytorch/pytorch/pull/144700