[inductor] Fix an aten.squeeze stride computation issue
Summary: Fixes https://github.com/pytorch/pytorch/issues/143498. The root cause is incorrect output stride for aten.squeeze (coming from aten.select in this case). If the input to aten.squeeze is non-contiguous, its output strides should also be non-contiguous. In addition to that, aten.cat also causes problem to the stride computation if it is turned into pointwise cat.

ghstack-source-id: 72cdeb7589b5298974242349313ccad65fe794f9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/143683